# Conflicts:
#	.gradle/8.0/executionHistory/executionHistory.bin
#	.gradle/8.0/executionHistory/executionHistory.lock
#	.gradle/8.0/fileHashes/fileHashes.bin
#	.gradle/8.0/fileHashes/fileHashes.lock
#	.gradle/8.0/fileHashes/resourceHashesCache.bin
#	.gradle/buildOutputCleanup/buildOutputCleanup.lock
#	app/build/intermediates/apk/debug/app-debug.apk
#	app/build/intermediates/dex/debug/mergeProjectDexDebug/7/classes.dex
#	app/build/intermediates/incremental/packageDebug/tmp/debug/dex-renamer-state.txt
#	app/build/intermediates/project_dex_archive/debug/out/com/example/eatraw/adapter/BestReviewAdapter.dex
#	app/build/kotlin/compileDebugKotlin/cacheable/caches-jvm/inputs/source-to-output.tab
#	app/build/kotlin/compileDebugKotlin/cacheable/caches-jvm/inputs/source-to-output.tab.values.at
#	app/build/kotlin/compileDebugKotlin/cacheable/caches-jvm/jvm/kotlin/class-attributes.tab
#	app/build/kotlin/compileDebugKotlin/cacheable/caches-jvm/jvm/kotlin/class-attributes.tab.values.at
#	app/build/kotlin/compileDebugKotlin/cacheable/caches-jvm/jvm/kotlin/class-fq-name-to-source.tab
#	app/build/kotlin/compileDebugKotlin/cacheable/caches-jvm/jvm/kotlin/class-fq-name-to-source.tab.values.at
#	app/build/kotlin/compileDebugKotlin/cacheable/caches-jvm/jvm/kotlin/internal-name-to-source.tab
#	app/build/kotlin/compileDebugKotlin/cacheable/caches-jvm/jvm/kotlin/internal-name-to-source.tab.values.at
#	app/build/kotlin/compileDebugKotlin/cacheable/caches-jvm/jvm/kotlin/proto.tab
#	app/build/kotlin/compileDebugKotlin/cacheable/caches-jvm/jvm/kotlin/proto.tab.values.at
#	app/build/kotlin/compileDebugKotlin/cacheable/caches-jvm/jvm/kotlin/source-to-classes.tab
#	app/build/kotlin/compileDebugKotlin/cacheable/caches-jvm/jvm/kotlin/source-to-classes.tab.values.at
#	app/build/kotlin/compileDebugKotlin/cacheable/caches-jvm/jvm/kotlin/subtypes.tab
#	app/build/kotlin/compileDebugKotlin/cacheable/caches-jvm/jvm/kotlin/subtypes.tab.values.at
#	app/build/kotlin/compileDebugKotlin/cacheable/caches-jvm/jvm/kotlin/supertypes.tab
#	app/build/kotlin/compileDebugKotlin/cacheable/caches-jvm/jvm/kotlin/supertypes.tab.values.at
#	app/build/kotlin/compileDebugKotlin/cacheable/caches-jvm/lookups/counters.tab
#	app/build/kotlin/compileDebugKotlin/cacheable/caches-jvm/lookups/file-to-id.tab
#	app/build/kotlin/compileDebugKotlin/cacheable/caches-jvm/lookups/file-to-id.tab.values.at
#	app/build/kotlin/compileDebugKotlin/cacheable/caches-jvm/lookups/id-to-file.tab
#	app/build/kotlin/compileDebugKotlin/cacheable/caches-jvm/lookups/id-to-file.tab.keystream
#	app/build/kotlin/compileDebugKotlin/cacheable/caches-jvm/lookups/id-to-file.tab.keystream.len
#	app/build/kotlin/compileDebugKotlin/cacheable/caches-jvm/lookups/id-to-file.tab.len
#	app/build/kotlin/compileDebugKotlin/cacheable/caches-jvm/lookups/id-to-file.tab.values.at
#	app/build/kotlin/compileDebugKotlin/cacheable/caches-jvm/lookups/id-to-file.tab_i
#	app/build/kotlin/compileDebugKotlin/cacheable/caches-jvm/lookups/lookups.tab
#	app/build/kotlin/compileDebugKotlin/cacheable/caches-jvm/lookups/lookups.tab.values
#	app/build/kotlin/compileDebugKotlin/cacheable/caches-jvm/lookups/lookups.tab.values.at
#	app/build/kotlin/compileDebugKotlin/cacheable/caches-jvm/lookups/lookups.tab.values.s
#	app/build/kotlin/compileDebugKotlin/cacheable/last-build.bin
#	app/build/kotlin/compileDebugKotlin/local-state/build-history.bin
#	app/build/tmp/kotlin-classes/debug/com/example/eatraw/adapter/BestReviewAdapter.class

+++ app/src/main/java/com/example/eatraw/ReviewActivity.kt
@@ -260,7 +260,7 @@ class ReviewActivity : AppCompatActivity() {
                     val storeImg = document["storeImg"] as String?
                     val region = document["region"] as String?
                     val like = (document["like"] as? Long)?.toInt() // "like" 필드를 Int로 가져오기
-                    val cost = (document["cost"] as? Double)?.toInt()
+                    val cost = (document["cost"] as? Long)?.toInt()
                     val fishKind = document["fishKind"] as String?
                     val userId = document["userId"] as String?
